Renamed "Globals" to "reactPolyfills".
