+++ src/index.js
@@ -1,4 +1,4 @@
-import 'Globals';
+import 'reactPolyfills';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import 'Objs/';
