+++ src/reactPolyfills.js
